requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Flask-Caching
python-dotenv
Werkzeug
#pandas
numpy
scikit-learn
scipy
matplotlib
seaborn
sentence-transformers
torch
transformers
surprise
SQLAlchemy
WTForms
email-validator
marshmallow
Jinja2
itsdangerous
gunicorn
click
joblib
requests
tmdbv3api